Adds support for single quotes in plan annotation
In setup seed node we use kubectl annotate to add the plan annotation
to the seed node.  The encoding is difficult as we need to escape the
string multiple times as we issue this call in a shell on the seed node.
This change removes the need to perform the escaping as it puts the
annotation in a file and copies it to the target machine.